refactor(mixin): move the accessor interfaces to their own package
The engine imports nine of the thirteen accessors from render/ and sodium/,
and those were the only names it may take from the mixin package: an
injecting mixin is never loaded as a class of its own, so a reference to
one from ordinary code fails at runtime and compiled while the two lived
side by side. With the accessors in mixin/access/ and nothing else there,
the package the rest of the engine sees is a leaf of pure interfaces, and
importing an injector becomes a compilation error.
