+++ common/src/main/java/dev/vitrail/mixin/access/ByteBufferBuilderAccessor.java
@@ -1,4 +1,4 @@
-package dev.vitrail.mixin;
+package dev.vitrail.mixin.access;
 
 import com.mojang.blaze3d.vertex.ByteBufferBuilder;
 import org.spongepowered.asm.mixin.Mixin;
