+++ common/src/main/java/dev/vitrail/render/GpuRecording.java
@@ -1,7 +1,7 @@
 package dev.vitrail.render;
 
-import dev.vitrail.mixin.CommandEncoderAccessor;
-import dev.vitrail.mixin.VulkanCommandEncoderAccessor;
+import dev.vitrail.mixin.access.CommandEncoderAccessor;
+import dev.vitrail.mixin.access.VulkanCommandEncoderAccessor;
 
 import com.mojang.blaze3d.systems.CommandEncoder;
 import com.mojang.blaze3d.systems.CommandEncoderBackend;
