+++ common/src/main/java/dev/vitrail/render/PackCompute.java
@@ -3,9 +3,9 @@ package dev.vitrail.render;
 import dev.vitrail.glsl.LoadClock;
 import dev.vitrail.glsl.PackProgram;
 import dev.vitrail.glsl.TranslatedUnit;
-import dev.vitrail.mixin.CommandEncoderAccessor;
-import dev.vitrail.mixin.GpuDeviceAccessor;
-import dev.vitrail.mixin.VulkanCommandEncoderAccessor;
+import dev.vitrail.mixin.access.CommandEncoderAccessor;
+import dev.vitrail.mixin.access.GpuDeviceAccessor;
+import dev.vitrail.mixin.access.VulkanCommandEncoderAccessor;
 import dev.vitrail.pack.program.ProgramNames;
 import dev.vitrail.pack.program.ProgramStage;
 import dev.vitrail.pack.program.RenderStage;
