+++ common/src/main/java/dev/vitrail/render/StorageBuffers.java
@@ -1,8 +1,8 @@
 package dev.vitrail.render;
 
-import dev.vitrail.mixin.CommandEncoderAccessor;
-import dev.vitrail.mixin.GpuDeviceAccessor;
-import dev.vitrail.mixin.VulkanCommandEncoderAccessor;
+import dev.vitrail.mixin.access.CommandEncoderAccessor;
+import dev.vitrail.mixin.access.GpuDeviceAccessor;
+import dev.vitrail.mixin.access.VulkanCommandEncoderAccessor;
 import dev.vitrail.pack.texture.BufferObject;
 import dev.vitrail.pack.texture.CustomStorage;
 import dev.vitrail.Vitrail;
