+++ common/src/main/resources/vitrail.mixins.json
@@ -8,26 +8,26 @@
 	},
 	"client": [
 		"ArenaAggregatorMixin",
-		"BiomeClimateAccessor",
+		"access.BiomeClimateAccessor",
 		"BiomeMixin",
 		"BiomesMixin",
 		"BlockEntityRenderDispatcherMixin",
-		"BlockEntityRenderStateAccessor",
+		"access.BlockEntityRenderStateAccessor",
 		"BlockRendererMixin",
 		"BufferBuilderMixin",
-		"ByteBufferBuilderAccessor",
+		"access.ByteBufferBuilderAccessor",
 		"ChunkMeshFormatsMixin",
 		"ChunkVertexMixin",
 		"ClientLevelMixin",
 		"CloudRendererMixin",
-		"CommandEncoderAccessor",
+		"access.CommandEncoderAccessor",
 		"CommandEncoderMixin",
 		"CustomFeatureRendererMixin",
 		"CustomSubmitMixin",
 		"DebugScreenEntriesMixin",
 		"DebugScreenEntryListMixin",
 		"GlslCompilerMixin",
-		"IntermediaryShaderModuleAccessor",
+		"access.IntermediaryShaderModuleAccessor",
 		"IntermediaryShaderModuleMixin",
 		"MinecraftStartupGuardMixin",
 		"MixinDefaultChunkRenderer",
@@ -39,7 +39,7 @@
 		"GameRendererMixin",
 		"GameRendererScaleMixin",
 		"GameRendererWarmupMixin",
-		"GpuDeviceAccessor",
+		"access.GpuDeviceAccessor",
 		"HudCompileIconMixin",
 		"ItemFeatureRendererMixin",
 		"ItemInHandRendererMixin",
@@ -49,7 +49,7 @@
 		"LevelExtractorMixin",
 		"LevelRendererFrameMixin",
 		"LevelRendererMixin",
-		"LevelRendererOutlineAccessor",
+		"access.LevelRendererOutlineAccessor",
 		"LevelRendererSkyMixin",
 		"ModelFeatureRendererMixin",
 		"ModelSubmitMixin",
@@ -57,23 +57,23 @@
 		"ProjectionMatrixBufferMixin",
 		"QuadParticleFeatureRendererMixin",
 		"RenderPassMixin",
-		"RenderPipelineAccessor",
+		"access.RenderPipelineAccessor",
 		"RenderPipelineMixin",
 		"MixinRenderRegion",
-		"RenderSectionManagerAccessor",
-		"RenderTargetAccessor",
+		"access.RenderSectionManagerAccessor",
+		"access.RenderTargetAccessor",
 		"RenderTypeFeatureRendererGroupMixin",
 		"RenderTypeFeatureRendererMixin",
 		"MixinShaderChunkRenderer",
 		"SkyRendererMixin",
-		"MixinSodiumWorldRenderer",
+		"access.MixinSodiumWorldRenderer",
 		"MixinSodiumWorldRendererInit",
 		"StagedVertexBufferDrawMixin",
-		"TextureManagerAccessor",
+		"access.TextureManagerAccessor",
 		"VKDrawContextMixin",
 		"VulkanBackendMixin",
 		"VulkanBindGroupLayoutMixin",
-		"VulkanCommandEncoderAccessor",
+		"access.VulkanCommandEncoderAccessor",
 		"VulkanCommandEncoderMixin",
 		"VulkanConstMixin",
 		"VulkanDeviceMixin",
